+++ sqlite3/lib/src/platform/fallback.dart
@@ -1,0 +1,5 @@
+import 'package:path/path.dart' as p;
+
+String pathToAbsoluteAndNormalize(String source) {
+  return p.url.normalize('/$source');
+}